+++ src/LineItemRender.cpp
@@ -24,7 +24,7 @@ void LineItemRender::render()
 {
     Render::render(m_pItem->m_matProj, m_pItem->m_matView, m_pItem->m_matModel, m_pItem->m_v4Color);
 
-    m_pShader->setVec2("v2Resolution", m_pItem->m_ptResolution.x, m_pItem->m_ptResolution.y); 
+    m_pShader->setVec2("v2Resolution", (float)m_pItem->m_ptResolution.x, (float)m_pItem->m_ptResolution.y); 
     m_pShader->setUInt("nPattern", m_pItem->m_nPattern);
     m_pShader->setFloat("dFactor", m_pItem->m_dFactor);
